E2E running of all metrics

+++ src/model_audit_cli/metrics/base_metric.py
@@ -7,7 +7,7 @@ class Metric:
     """Base class for all metrics."""
 
     name: str
-    value: Union[float, Dict[str, int]]
+    value: Union[float, Dict[str, float]]
     latency_ms: int
     details: Dict[str, Any]
